+++ packages/fiori-app-sub-generator/src/fiori-app-generator/transforms.ts
@@ -220,7 +220,8 @@ export async function transformState<T>(
                 project.skipAnnotations !== true
                     ? await getAnnotations(project.name, service.annotations?.[0], service?.capService)
                     : undefined,
-            ignoreCertError: service.ignoreCertError
+            ignoreCertError: service.ignoreCertError,
+            externalServices: service.valueListMetadata
         };
 
         const destinationName = service.destinationName ?? service.connectedSystem?.destination?.Name;