+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.69
+
+### Patch Changes
+
+- Updated dependencies [691ab27]
+    - @sap-ux/odata-service-inquirer@2.13.0
+    - @sap-ux/inquirer-common@0.10.1
+    - @sap-ux/abap-deploy-config-inquirer@1.6.63
+
 ## 0.2.68
 
 ### Patch Changes